Initial Authentication Flow
This adds the basic code for the initial authentication flow.  It
includes spawning a server to retrieve the OAuth credentials returned
in GET parameters, sending the user off to the OAuth endpoint, and
fetching the OAuth tokens once the user has accepted.

This is a little bit longer than ideal, but it also includes some
helpers that will make future auth/API code significantly shorter.