Added Rate Limiting Support
My initial plan was to just process in small enough batches to avoid
rate limiting, but it turns out the Reddit API won't let you skip to
the last page of results, which means I'm going to have to paginate
all the way to the oldest posts for the user which means a
conservative number of requests may not be enough.  So this adds some
code to the request helpers to ensure that we pause if we're about to
hit the rate limit.